Add a name and description to spots in maps.me file
I had to fix the encoding in the database. Here is what I did manually in IRB for each record:
- symbolize keys
- set :id as :hw_id
- remove :_id
- apply the code of Spot#fix_encoding on this modified hash
- pass this hash to Spot.new
- save within a begin/rescue block that rescues ArgumentError (as it can be raised for invalid UTF-8)

This had to be made before merging this commit as now Spot#save does the encoding fix.

The problem was that MAPS.me would ignore an entire collection of pins if the encoding is not good. And for some reason, it did not work to fix the encoding on the fly while building the file. And I could not find how to solve it, so I decided to fix the encoding upstream.