Better API for DB interraction
It bothered be to have a mix of raw and modified data in a document.
Plus it seems better to me to separate the DB interraction and the Hitchspot::Spot api whose data storage could change.

With this structure, document#raw is just the hitchwiki raw data, and document#sanitized is the sanitized data that we interract with in the app. This way, sanitized can be modified safely without modifying the raw data who stays as a reference

New rake file: data migrations, should be ran only once, in order