More data in maps.me
Now that we have nice sanitized data, let's add it to the template. As explained here https://developers.google.com/kml/documentation/kml_tut it's possible to have an html description if it is properly escaped.

Added: average waiting time, comments with author and date, author and date of description, titles, style